+++ websitebuild-20250429/alibabacloud_websitebuild20250429/models/_report_channel_publish_result_request.py
@@ -7,23 +7,26 @@ from darabonba.model import DaraModel
 class ReportChannelPublishResultRequest(DaraModel):
     def __init__(
         self,
+        channel_account_name: str = None,
         draft_id: str = None,
         external_id: str = None,
         external_url: str = None,
         fail_reason: str = None,
         success: bool = None,
     ):
+        # The display name of the publishing account. A null value does not overwrite the original value. Maximum length: 256 characters.
+        self.channel_account_name = channel_account_name
         # The channel draft ID.
         # 
         # This parameter is required.
         self.draft_id = draft_id
-        # The ID returned by the platform. Specify this parameter when the publish operation is successful.
+        # The ID returned by the platform. Set this parameter when the publishing is successful.
         self.external_id = external_id
         # The redirect URL of the platform.
         self.external_url = external_url
-        # The failure reason. Specify this parameter when the publish operation fails.
+        # The failure reason. Set this parameter when the publishing fails.
         self.fail_reason = fail_reason
-        # Specifies whether the publish operation is successful (true/false).
+        # Specifies whether the publishing is successful (true/false).
         # 
         # This parameter is required.
         self.success = success
@@ -36,6 +39,9 @@ class ReportChannelPublishResultRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.channel_account_name is not None:
+            result['ChannelAccountName'] = self.channel_account_name
+
         if self.draft_id is not None:
             result['DraftId'] = self.draft_id
 
@@ -55,6 +61,9 @@ class ReportChannelPublishResultRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('ChannelAccountName') is not None:
+            self.channel_account_name = m.get('ChannelAccountName')
+
         if m.get('DraftId') is not None:
             self.draft_id = m.get('DraftId')