+++ websitebuild-20250429/alibabacloud_websitebuild20250429/models/_rewrite_channel_title_request.py
@@ -14,7 +14,7 @@ class RewriteChannelTitleRequest(DaraModel):
         # 
         # This parameter is required.
         self.draft_id = draft_id
-        # The user personalization requirement in free text format. This parameter is optional.
+        # The user personalization requirement in free text. This parameter is optional.
         self.user_requirement = user_requirement
 
     def validate(self):