+++ websitebuild-20250429/alibabacloud_websitebuild20250429/models/_save_channel_draft_response_body.py
@@ -29,8 +29,8 @@ class SaveChannelDraftResponseBody(DaraModel):
         self.app_name = app_name
         # The dynamic error code.
         self.dynamic_code = dynamic_code
-        # The dynamic error message, which is used to replace the `%s` in the **ErrMessage** return parameter.
-        # > If **ErrMessage** returns **The Value of Input Parameter %s is not valid** and **DynamicMessage** returns **DtsJobId**, the request parameter **DtsJobId** that you specify is invalid.
+        # The dynamic error message, which is used to replace the `%s` placeholder in the **ErrMessage** return parameter.
+        # > For example, if **ErrMessage** returns **The Value of Input Parameter %s is not valid** and **DynamicMessage** returns **DtsJobId**, the request parameter **DtsJobId** is invalid.
         self.dynamic_message = dynamic_message
         # The error parameters returned.
         self.error_args = error_args